+++ vim-engine/src/main/kotlin/com/maddyhome/idea/vim/action/motion/leftright/MotionMiddleColumnAction.kt
@@ -5,7 +5,7 @@
  * license that can be found in the LICENSE.txt file or at
  * https://opensource.org/licenses/MIT.
  */
-package com.maddyhome.idea.vim.action.macro
+package com.maddyhome.idea.vim.action.motion.leftright
 
 import com.maddyhome.idea.vim.api.ExecutionContext
 import com.maddyhome.idea.vim.api.VimCaret
